pyproject.toml
[project]
name = "cmt-thesis"
version = "0.1.0"
description = "CBS Master Thesis: Manifold AI Research Framework + Production System"
requires-python = ">=3.11"

dependencies = [
    "pyzotero>=1.5.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "langgraph>=0.2.0",
    "langchain>=0.1.0",
]
